# apidocs.yml
openapi: 3.0.0
info:
  title: My Basket Application
  description: >
    <div>Intro - MyBasket flask restful API
    <h1>Use:</h1>
    <ol> 
      <li>CRUD on Users</li>
      <li>CRUD on Products</li>
      <li>CRUD on Categories</li>
      <li>CRUD on Manager Requests</li>
      <li>CRUD on Feedback</li>
      <li> Add Items to Cart and update quantity</li>
      <li>Place Orders</li>
      <li>Search a Product</li>
    </ol>
    </div>

    <h1> Users Table Schema </h1>
    <table>
      <thead>
        <th>Column Name</th>
        <th>Column Type</th>
        <th>Constraints</th>
      </thead>
      <tbody>
        <tr>
          <td>id</td>
          <td>Integer</td>
          <td>Primary Key, Auto Increment</td>
        </tr>
        <tr>
          <td>username</td>
          <td>String</td>
          <td>Unique, Not Null</td>
        </tr>
        <tr>
          <td>password</td>
          <td>String</td>
          <td>Not Null</td>
        </tr>
        <tr>
          <td>email</td>
          <td>String</td>
          <td>Not Null</td>
        </tr>
        <tr>
          <td>delivery_details</td>
          <td>Text</td>
          <td></td>
        </tr>
        <tr>
          <td>created_at</td>
          <td>DateTime</td>
          <td>Default: datetime.utcnow, Not Null</td>
        </tr>
        <tr>
          <td>last_activity</td>
          <td>DateTime</td>
          <td>Default: datetime.utcnow, Not Null</td>
        </tr>
        <tr>
...
        </tr>
...

...

...


...

...
servers:
...
paths:
...
  /user/{user_id}:
    get:
      tags:
        - Users
      summary: Get user details by user ID
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: integer
          description: The ID of the user to retrieve.
      responses:
        '200':
          description: User details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '403':
          description: Permission denied
          content:
            application/json:
              example:
                message: Permission denied
        '404':
          description: User not found
          content:
            application/json:
              example:
                message: User not found
      security:
        - jwt: []
...
components:
  securitySchemes:
    jwt:
      type: apiKey
      in: header
      name: Authorization
  schemas:
...
     User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          example: dummy
        email:
          type: string
          example: dummy@yopmail.com
        created_at:
          type: string
          format: date-time
          example: '2022-12-29T14:27:45Z'
        approved:
          type: boolean
          example: true
        active:
          type: boolean
          example: true
        role_id:
          type: integer
          example: 2
        report_type:
          type: string
          example: general
        following_count:
          type: integer
          example: 0
        post_count:
          type: integer
          example: 0
        account_created:
          type: string
          example: Thu, 29 Dec 2022 14:27:45 -0000
        image_file:
          type: string
          example: [image-url]
...